refactor: modularize backend and add docs
